pyproject.toml
[tool.poetry]
name = "slope-stability-forecast"
version = "1.0.0"
description = "Code to fetch field monitoring data along with meteorlogical conditions to forecast factor of safety of a slope and display the results in NGI Live"
authors = [""]
readme = "README.md"

[tool.poetry.dependencies]
keras= "^2.6.0"
python = "^3.11"
numpy = "^1.26.4"
pandas = "^2.2.1"
requests = "^2.31.0"
scikit-learn = "1.3.1"
joblib = "^1.3.2"
python-dotenv = "^1.0.1"
azure-storage-blob = "^12.19.0"
azure-keyvault-secrets = "^4.8.0"
azure-identity = "^1.15.0"
pastas = "^1.4.0"
statsmodels = "^0.14.1"
pmdarima = "^2.0.4"
absl-py = "2.0.0"
datetime = "^5.4"
openmeteo-requests = "^1.2.0"
requests-cache = "^1.2.0"
retry-requests = "^2.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
